Rename oci-pilot -> podman-pilot
This Fixes #60
