--- oci-pilot_test/usr/sbin/apt-get
@@ -1,1 +1,0 @@
-../../../oci-pilot/target/debug/oci-pilot
