--- oci-pilot_test/usr/sbin/python3
@@ -1,1 +1,0 @@
-../../../oci-pilot/target/debug/oci-pilot
