--- oci-pilot_test/usr/sbin/salt
@@ -1,1 +1,0 @@
-../../../oci-pilot/target/debug/oci-pilot
