+++ podman-pilot/Cargo.toml
@@ -6,7 +6,7 @@ codegen-units = 1
 panic = "abort"
 
 [package]
-name = "oci-pilot"
+name = "podman-pilot"
 version = "2.0.5"
 edition = "2018"
